feat(global_rec): implement background indexer, refactor and modularize global rec implementation.
